refactor(client): reorg tests
